Move theme.js
