+++ src/pages/build.js
@@ -7,7 +7,7 @@ import PageMetadata from "../components/PageMetadata"
 import Translation from "../components/Translation"
 import ButtonLink from "../components/ButtonLink"
 import Link from "../components/Link"
-import { Mixins } from "../components/Theme"
+import { Mixins } from "../theme"
 import ActionCard from "../components/ActionCard"
 import { Divider } from "../components/SharedStyledComponents"
 import Emoji from "../components/Emoji"
